Add files via upload
A major issue that has an impact on communities all across the world is road safety. In this project I have created a Road Accident Dashboard in Power BI using data for the years 2021 and 2022. This dashboard empowers and serves as a valuable tool for stakeholders such as Ministry of Transport, Traffic Management Agencies, Road Safety Corps and Road Transport Department with valuable insights to make data-driven decisions and implement evidence-based interventions that enhance road safety. 
Features and Functionalities you will learn include; 
•	Using Advanced DAX functions to create Measures 
•	How to group in Power BI
•	Using Advanced KPI
•	Using Filter